Add files via upload
Created a python script that allows you to convert your bank statements and rename them and move them to a folder. very helpful if you have multiple ones. Used a test to run on a terminal to parse through the documents and make sure it was capturing the right information, then using that information pulled the information out. 

This is strictly for PNC but you can use it for your own bank and modify and tweak it.